create models folder, move relevant files, update imports
